% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agencies_initials.R
\docType{data}
\name{agencies_initials}
\alias{agencies_initials}
\title{Public agencies available.}
\format{
A data frame with 311 rows and 2 variables:
\describe{
\item{agencies_initials}{public agency initials}
\item{agencies_names}{public agency name}

}
}
\source{
\url{http://www.consultaesic.cgu.gov.br/busca/_layouts/15/downloadpedidos/downloaddados.aspx}
}
\usage{
agencies_initials
}
\description{
A dataset containing federal public agencies available to search in the agency argument from the functions requests and appeals.
}
\keyword{datasets}
