% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeals_all.R
\name{appeals_all}
\alias{appeals_all}
\title{Query the appeals made through Right to Information Law to Brazilian federal government}
\usage{
appeals_all(year = "all")
}
\arguments{
\item{year}{selects which years data will be downloaded}
}
\value{
a dataframe with requests to access information containing the keyword
}
\description{
Download data from the CGU for the selected years, apply a filter and return it in the form of a dataframe.
}
\examples{
\dontrun{appeals_all()}
}
