% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests_all.R
\name{requests_all}
\alias{requests_all}
\title{Query all the requests made through Right to Information Law to Brazilian federal government
Download data from the CGU for the selected years and return it in the form of a dataframe.}
\usage{
requests_all(year = "all")
}
\arguments{
\item{year}{selects which years data will be downloaded}
}
\value{
a dataframe with requests to access information containing the keyword
}
\description{
Query all the requests made through Right to Information Law to Brazilian federal government
Download data from the CGU for the selected years and return it in the form of a dataframe.
}
\examples{
\dontrun{requests_all()}
}
