% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeals.R
\name{appeals}
\alias{appeals}
\title{Query the appeals made through Right to Information Law to Brazilian federal government}
\usage{
appeals(year = "all", search)
}
\arguments{
\item{year}{selects which years data will be downloaded}

\item{search}{select the keyword to be searched}
}
\value{
a dataframe with appeals made to responses given to requests for information containing the keyword
}
\description{
Download data from the CGU for the selected years, apply a filter and return it in the form of a dataframe.
}
\examples{
\dontrun{appeals(search = 'PAC')}
}
