% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeals.R
\name{appeals}
\alias{appeals}
\title{Queries the appeals made through the Brazilian Right to Information Law}
\usage{
appeals(year = "all", answer = F, search = "all")
}
\arguments{
\item{year}{selects the years which data will be downloaded}

\item{answer}{if true, fetches the content of the search argument in the appeals responses}

\item{search}{selects the keyword to be searched}
}
\value{
a dataframe with appeals containing the keyword
}
\description{
Downloads data for the selected years, apply a filter and return it in the form of a dataframe.
}
\examples{
\dontrun{appeals(search = 'PAC')}
}
