% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeals_all.R
\name{appeals_all}
\alias{appeals_all}
\title{Queries all the appeals made through the Brazilian Right to Information Law}
\usage{
appeals_all(year = "all")
}
\arguments{
\item{year}{selects which years data will be downloaded}
}
\value{
a dataframe with all appeals made
}
\description{
Downloads all the data and return it in the form of a dataframe.
}
\examples{
\dontrun{appeals_all()}
}
