% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{requests}
\alias{requests}
\title{Query the requests made through Right to Information Law to Brazilian federal government}
\usage{
requests(year = "all", search)
}
\arguments{
\item{year}{selects which years data will be downloaded}

\item{search}{select the keyword to be searched}
}
\value{
a dataframe with requests to access information containing the keyword
}
\description{
Download data from the CGU for the selected years, apply a filter and return the data in the form of a dataframe
}
\examples{
\dontrun{requests(search = 'PAC')}
}
