% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{requests}
\alias{requests}
\title{Queries the requests made through the Brazilian Right to Information Law}
\usage{
requests(year = "all", answer = F, search = "all")
}
\arguments{
\item{year}{selects the years which data will be downloaded}

\item{answer}{if true, fetches the content of the search argument in the request responses}

\item{search}{selects the keyword to be searched}
}
\value{
a dataframe with requests containing the keyword
}
\description{
Downloads data for the selected years, apply a filter and return it in the form of a dataframe.
}
\examples{
\dontrun{requests(search = 'PAC')}
}
