% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests_all.R
\name{requests_all}
\alias{requests_all}
\title{Queries all the requests made through the Brazilian Right to Information Law}
\usage{
requests_all(year = "all")
}
\arguments{
\item{year}{selects which years data will be downloaded}
}
\value{
a dataframe with all requests made
}
\description{
Downloads all the data and return it in the form of a dataframe.
}
\examples{
\dontrun{requests_all()}
}
